First version working and ready for K8s
The first version of BadStats is working and ready to be deployed
as a containerized application on a K8s cluster.

This release has a minimal feature set but has basic artist, album, and
song lookup features. It can display all of the basic aspects of each
type of item. The big feature is the album plot, which shows the
distribution of a song characteristic throughout an album.