Updated README.md and housekeeping
README.md updated to better reflect current state of the app. Other
housekeeping was done as well, such as renaming requirements-prod.txt
to simply requirements.txt. As a result of the requirements.txt rename,
the Dockerfile had to be updated. The .dockerignore was updated as well.

The start.sh script was removed because it wasn't being used often.